Bugfixes:thermite, zombie FF, river slow, drinkingglass, kitchen tray, tank reload, full window, etc.

Closes #5218, #6368, #6062, #5545, #5617, #5570, #5422, #6305, #6989, and #6992

See merge request cmdevs/ColonialMarines!3752

+++ code/game/turfs/walls/walls.dm
@@ -232,13 +232,15 @@
 
 		sleep(20)
 		if(!istype(src, /turf/closed/wall)) // Extra check, needed against runtimes
-			break
+			if(O && !O.disposed)
+				qdel(O)
+			return
 	melting = FALSE
 	if(destroyed)
 		dismantle_wall(1)
-	if(O)
+	if(O && !O.disposed)
 		qdel(O)
-	return
+
 
 
 //Interactions
